RF: Move custom argparse formatters away (THERE WILL BE NO TESTS FOR THOSE!)
